Fix 32-bit build after 530d274c15e5.
MFC after:	3 days

(cherry picked from commit 05c3e8e87142a8689503a6731237b48fb43c1bec)

+++ sys/cam/ctl/ctl.c
@@ -1104,8 +1104,8 @@ ctl_isc_ua(struct ctl_softc *softc, union ctl_ha_msg *msg, int len)
 	uint32_t iid;
 
 	if (len < sizeof(msg->ua)) {
-		printf("%s: Received truncated message %d < %lu\n",
-		    __func__, len, sizeof(msg->ua));
+		printf("%s: Received truncated message %d < %d\n",
+		    __func__, len, (int)sizeof(msg->ua));
 		ctl_ha_msg_abort(CTL_HA_CHAN_CTL);
 		return;
 	}
@@ -1145,15 +1145,15 @@ ctl_isc_lun_sync(struct ctl_softc *softc, union ctl_ha_msg *msg, int len)
 	uint32_t targ_lun;
 
 	if (len < offsetof(struct ctl_ha_msg_lun, data[0])) {
-		printf("%s: Received truncated message %d < %lu\n",
-		    __func__, len, offsetof(struct ctl_ha_msg_lun, data[0]));
+		printf("%s: Received truncated message %d < %d\n",
+		    __func__, len, (int)offsetof(struct ctl_ha_msg_lun, data[0]));
 		ctl_ha_msg_abort(CTL_HA_CHAN_CTL);
 		return;
 	}
 	i = msg->lun.lun_devid_len + msg->lun.pr_key_count * sizeof(pr_key);
 	if (len < offsetof(struct ctl_ha_msg_lun, data[i])) {
-		printf("%s: Received truncated message data %d < %lu\n",
-		    __func__, len, offsetof(struct ctl_ha_msg_lun, data[i]));
+		printf("%s: Received truncated message data %d < %d\n",
+		    __func__, len, (int)offsetof(struct ctl_ha_msg_lun, data[i]));
 		ctl_ha_msg_abort(CTL_HA_CHAN_CTL);
 		return;
 	}
@@ -1229,8 +1229,8 @@ ctl_isc_port_sync(struct ctl_softc *softc, union ctl_ha_msg *msg, int len)
 	int i, new;
 
 	if (len < offsetof(struct ctl_ha_msg_port, data[0])) {
-		printf("%s: Received truncated message %d < %lu\n",
-		    __func__, len, offsetof(struct ctl_ha_msg_port, data[0]));
+		printf("%s: Received truncated message %d < %d\n",
+		    __func__, len, (int)offsetof(struct ctl_ha_msg_port, data[0]));
 		ctl_ha_msg_abort(CTL_HA_CHAN_CTL);
 		return;
 	}
@@ -1238,8 +1238,8 @@ ctl_isc_port_sync(struct ctl_softc *softc, union ctl_ha_msg *msg, int len)
 	    msg->port.port_devid_len + msg->port.target_devid_len +
 	    msg->port.init_devid_len;
 	if (len < offsetof(struct ctl_ha_msg_port, data[i])) {
-		printf("%s: Received truncated message data %d < %lu\n",
-		    __func__, len, offsetof(struct ctl_ha_msg_port, data[i]));
+		printf("%s: Received truncated message data %d < %d\n",
+		    __func__, len, (int)offsetof(struct ctl_ha_msg_port, data[i]));
 		ctl_ha_msg_abort(CTL_HA_CHAN_CTL);
 		return;
 	}
@@ -1358,15 +1358,15 @@ ctl_isc_iid_sync(struct ctl_softc *softc, union ctl_ha_msg *msg, int len)
 	int i, iid;
 
 	if (len < offsetof(struct ctl_ha_msg_iid, data[0])) {
-		printf("%s: Received truncated message %d < %lu\n",
-		    __func__, len, offsetof(struct ctl_ha_msg_iid, data[0]));
+		printf("%s: Received truncated message %d < %d\n",
+		    __func__, len, (int)offsetof(struct ctl_ha_msg_iid, data[0]));
 		ctl_ha_msg_abort(CTL_HA_CHAN_CTL);
 		return;
 	}
 	i = msg->iid.name_len;
 	if (len < offsetof(struct ctl_ha_msg_iid, data[i])) {
-		printf("%s: Received truncated message data %d < %lu\n",
-		    __func__, len, offsetof(struct ctl_ha_msg_iid, data[i]));
+		printf("%s: Received truncated message data %d < %d\n",
+		    __func__, len, (int)offsetof(struct ctl_ha_msg_iid, data[i]));
 		ctl_ha_msg_abort(CTL_HA_CHAN_CTL);
 		return;
 	}
@@ -1396,8 +1396,8 @@ ctl_isc_login(struct ctl_softc *softc, union ctl_ha_msg *msg, int len)
 {
 
 	if (len < sizeof(msg->login)) {
-		printf("%s: Received truncated message %d < %lu\n",
-		    __func__, len, sizeof(msg->login));
+		printf("%s: Received truncated message %d < %d\n",
+		    __func__, len, (int)sizeof(msg->login));
 		ctl_ha_msg_abort(CTL_HA_CHAN_CTL);
 		return;
 	}
@@ -1436,15 +1436,15 @@ ctl_isc_mode_sync(struct ctl_softc *softc, union ctl_ha_msg *msg, int len)
 	uint32_t initidx, targ_lun;
 
 	if (len < offsetof(struct ctl_ha_msg_mode, data[0])) {
-		printf("%s: Received truncated message %d < %lu\n",
-		    __func__, len, offsetof(struct ctl_ha_msg_mode, data[0]));
+		printf("%s: Received truncated message %d < %d\n",
+		    __func__, len, (int)offsetof(struct ctl_ha_msg_mode, data[0]));
 		ctl_ha_msg_abort(CTL_HA_CHAN_CTL);
 		return;
 	}
 	i = msg->mode.page_len;
 	if (len < offsetof(struct ctl_ha_msg_mode, data[i])) {
-		printf("%s: Received truncated message data %d < %lu\n",
-		    __func__, len, offsetof(struct ctl_ha_msg_mode, data[i]));
+		printf("%s: Received truncated message data %d < %d\n",
+		    __func__, len, (int)offsetof(struct ctl_ha_msg_mode, data[i]));
 		ctl_ha_msg_abort(CTL_HA_CHAN_CTL);
 		return;
 	}
